update types and remove optionalPackages
