+++ package.json
@@ -11,20 +11,17 @@
     "node": ">=16"
   },
   "scripts": {
-    "start": "npm run copy-dev & yarn watch",
+    "start": "npm run watch",
     "build:types": "tsc --project tsconfig.json",
-    "build": "yarn copy-prod & rimraf dist & node esbuild.js & yarn build:types",
-    "watch": "yarn copy-prod & rimraf dist & yarn build:types & node esbuild-watch.mjs",
+    "build": "rimraf dist & node esbuild.mjs & yarn build:types",
+    "watch": "rimraf dist & yarn build:types & node esbuild-watch.mjs",
     "test": "dts test --passWithNoTests",
     "lint": "dts lint",
     "prepare": "npm run build",
     "size": "size-limit",
-    "postinstall": "yarn copy-prod",
     "analyze": "size-limit --why",
     "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook",
-    "copy-dev": "shx cp src/optionalPackages/dev/* src/optionalPackages",
-    "copy-prod": "shx cp src/optionalPackages/prod/* src/optionalPackages"
+    "build-storybook": "build-storybook"
   },
   "devDependencies": {
     "@babel/core": "7.17.8",
@@ -49,7 +46,6 @@
     "dts-cli": "^1.5.1",
     "esbuild": "^0.14.43",
     "esbuild-node-externals": "^1.4.1",
-    "esbuild-plugin-globals": "^0.1.1",
     "esbuild-plugin-svgr": "^1.0.1",
     "eslint-config-prettier": "8.5.0",
     "eslint-config-react": "1.1.7",
@@ -65,7 +61,6 @@
     "eslint-plugin-react-hooks": "4.5.0",
     "husky": "8.0.1",
     "jest": "28.1.1",
-    "node-stdlib-browser": "^1.2.0",
     "react": "18.1.0",
     "react-dom": "18.1.0",
     "react-is": "18.0.0",
@@ -74,7 +69,6 @@
     "rollup-plugin-svg-import": "1.6.0",
     "sass": "1.32",
     "size-limit": "7.0.8",
-    "stream-browserify": "^3.0.0",
     "ts-jest": "28.0.4",
     "tsdx": "0.14.1",
     "tslib": "2.4.0",
@@ -104,6 +98,7 @@
     "@fortawesome/fontawesome-svg-core": "6.1.0",
     "@fortawesome/free-solid-svg-icons": "6.1.0",
     "@fortawesome/react-fontawesome": "0.1.18",
+    "bootstrap": "4.6.0",
     "classnames": "2.3.1",
     "moment": "2.29.1",
     "platform": "1.3.6",
@@ -134,10 +129,10 @@
     "@elrondnetwork/erdjs-wallet-connect-provider": "2.0.2",
     "@elrondnetwork/erdjs-walletcore": "1.0.0",
     "@elrondnetwork/erdjs-web-wallet-provider": "2.1.0",
-    "@reduxjs/toolkit": "^1.8.2",
+    "@reduxjs/toolkit": "1.8.2",
     "autoprefixer": "9.8.8",
     "axios": "0.24.0",
-    "bignumber.js": "9.0.1",
+    "bignumber.js": "9.x",
     "lodash.debounce": "4.0.8",
     "lodash.isequal": "4.5.0",
     "lodash.omit": "4.5.0",
@@ -145,7 +140,7 @@
     "qs": "6.10.3",
     "react-error-overlay": "6.0.9",
     "react-idle-timer": "5.0.0",
-    "react-redux": "8.0.1",
+    "react-redux": "8.0.2",
     "redux-persist": "6.0.0",
     "reselect": "4.0.0",
     "rollup-plugin-peer-deps-external": "2.2.4",
