+++ src/constants/index.ts
@@ -1,6 +1,6 @@
-export * as errorsMessages from './errorsMessages';
-export * as networkConstants from './network';
-export * as ledgerErrorCodes from './ledgerErrorCodes';
+export * from './errorsMessages';
+export * from './network';
+export * from './ledgerErrorCodes';
 export * from './mnemonicWords';
 
 export const gasPriceModifier = '0.01';
