+++ src/hooks/index.tsx
@@ -4,4 +4,3 @@ export * from './useGetNotification';
 export * from './transactions';
 export * from './account';
 export * from './useDebounce';
-export * from './useIdleTimer';
