--- src/optionalPackages/dev/react-bootstrap.tsx
@@ -1,3 +1,0 @@
-import * as RB from 'react-bootstrap';
-
-export default RB;
