--- src/optionalPackages/prod/react-bootstrap.tsx
@@ -1,6 +1,0 @@
-let ReactBootstrap: any = {};
-try {
-  ReactBootstrap = require('react-bootstrap');
-} catch (err) {}
-
-export default ReactBootstrap;
