+++ src/types/UI.ts
@@ -1,0 +1,10 @@
+export interface DenominateType {
+  value: string;
+  showLastNonZeroDecimal?: boolean;
+  showLabel?: boolean;
+  token?: string;
+  decimals?: number;
+  denomination?: number;
+  egldLabel?: string;
+  'data-testid'?: string;
+}
