+++ src/types/index.ts
@@ -58,3 +58,4 @@ export interface TokenLoginType {
 export * from './transactions';
 export * from './enums';
 export * from './dapp-provider';
+export * from './UI';
